r2|r4 refactoring

+++ src/Layers/xrRenderDX10/dx11Texture.cpp
@@ -162,7 +162,7 @@ ID3DBaseTexture* CRender::texture_load(LPCSTR fRName, u32& ret_msize, bool bStag
 
 		if (pTexture2D == nullptr)
 		{
-			auto isPowerOfTwo = [](int x)
+			auto isPowerOfTwo = [](size_t x)
 			{
 				return (x && !(x & (x - 1)));
 			};
@@ -207,7 +207,7 @@ ID3DBaseTexture* CRender::texture_load(LPCSTR fRName, u32& ret_msize, bool bStag
 		
 		if (pTexture2D == nullptr)
 		{
-			auto isPowerOfTwo = [](int x)
+			auto isPowerOfTwo = [](size_t x)
 			{
 				return (x && !(x & (x - 1)));
 			};
